* Suse updated to spec and init files [Shinichi Furuso]
I tested compiling  on sles10, sles11, and opensuse 11.4.  I use
opensuse just for testing.

I'm using 'build' package (http://gitorious.org/opensuse/build), which
is a wrapper of rpmbuild, compiles rpm in chroot'ed environment and runs
rpmlint on it.

rc.patch modifies Required-Start tag. It is xl2tpd runs after $remote_fs
since it resides on /usr partition.

xl2tpd.spec is for sles11 and opensuse 11.4.  I decided not to share it
with sles10, because BuildRequires differs.  Of course, it can be merged
using 'if %{sles_version}', but even on sles11, sles_version is set to 0
in the 'build' package.  It doesn't harm since it can be compiled with
the same environment of opensuse, but I think it is misleading that
'sles_version is 0' means it is sles11 or opensuse 10/11.

sles10.spec is xl2tpd.spec for sles10.

Signed-off-by: Paul Wouters <paul@xelerance.com>