Don't call grantpt().
Use of grantpt() is fraught: we run with a SIGCHLD handler, which
can lead to undefined results on systems that exec a setuid child
process to fiddle the ownership/permissions of the slave pty when a
child pppd process dies at just the right moment.  In addition, it
reportedly interacts badly with SELinux.

It turns out that there's no need to call grantpt() anyway.  Child pppd
processes run as root and so should be able to access the slave ptys with
their original ownership and permissions.  On FreeBSD/Mac OS X, grantpt()
doesn't actually do anything interesting and has been presumably been
working fine.  Fedora has been shipping with a patch to disable the call
to grantpt, and OpenBSD is completely unaffected, so this looks pretty
safe.