Fix build errors caused by inline function with gcc 5
GCC 5 defaults to -std=gnu11 instead of -std=gnu89. And -std=gnu89
employs the GNU89 inline semantics, -std=gnu11 uses the C99 inline
semantics.

For 'inline' fuction, it is NOT exported by C99. So error messages such as:

| control.c:1717: undefined reference to `check_control'

For these functions which is not referred by other compile units, make
them 'static inline'.

For 'extern inline' function, it fails such as:

| misc.h:68:20: warning: inline function 'swaps' declared but never defined
|  extern inline void swaps (void *, int);
|                      ^

Because function swaps() is referred by other compile units, it must be
exported. The semantics of 'extern inline' are not same between GNU89
and C99, so remove 'inline' attribute for compatible with GNU89.

Ref:
https://gcc.gnu.org/gcc-5/porting_to.html

Signed-off-by: Kai Kang <kai.kang@windriver.com>