Remove unnecessary casting of malloc() results
AIUI in C, code like

    ptr = (char *)malloc(30);

should be written as

    ptr = malloc(30);

as the casting can lead to problems if you forget to include stdlib.h
due to implicitly declared functions having a return type of int. Though
these days GCC will warn loudly about any implicitly declared functions.
See here[0] for a lengthy discussion on the subject, more so the
comments.

There are it seems arguments both for casting as well as not to.
Personally I don't and most C code I read doesn't. In the case of xl2tpd
there are currently 17 calls to malloc and of those, 10 cast the result
and 7 don't.

[0] -
https://www.securecoding.cert.org/confluence/display/c/MEM02-C.+Immediately+cast+the+result+of+a+memory+allocation+function+call+into+a+pointer+to+the+allocated+type?src=contextnavpagetreemode

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>