avp: Early failure if no handler to remove indent block
Like most of te function code above, first check the failure case and
early return with error (or jump to next avp). This way the changed code
block is more similar to the other blocks of the function and we remove
anindentation level for free.

This commit was proposed and contributued by github user: pespin (Pau Espin Pedrol)