Sockopt bug fix for multiple IP's
Allow the use of IPsec SAref if force userspace is enabled.

There's a bug with xl2tpd where if it's unable to set the socket option for IPsec SAREF processing in the kernel, it'll never reach the "else" condition to set the sock opt for userspace ipsec saref.

This only appears to affect setups where "listen-addr" is set to 0.0.0.0 (all IP's)

So to remedy this, check if "force userspace" is enabled in xl2tpd.conf, If it is, don't bother checking for kernel support, continue on and set the sockopt for userspace.