debian: Add systemd .service
This is intended to be functionally equivalent to the existing init
script at debian/xl2tpd.init, and in particular to respect any changes
made to /etc/default/xl2tpd on existing installations. As a result,
rather than using RuntimeDirectory=xl2tpd to create /var/run/xl2tpd, it
is created explicitly with an ExecStartPre command.

Fixes #50 (and #51: no tmpfiles snippet is needed since the directory is
created when the unit is activated).