Ditch select() system call in favor of poll().
This moves from select() to poll().

There are potentially more improvements that can be made (not to rebuild
the entire pollfd array every time).  I believe a migration to epoll()
may also be of great benefit, however, this would require a reasonable
re-architecture, specifically, the event logic would need two
implementations:

1.  A POSIX compliant poll() bassed implementation; and
2.  A Linux specific epoll() implementation.

This should be possible, but I'm unsure if any further work is worth the
effort at this stage.

Bug: #276
Signed-off-by: Jaco Kroon <jaco@uls.co.za>